+++ src/test/BitPackGPU_test.cpp
@@ -94,9 +94,9 @@ template <typename T>
 void runBitPackingOnGPU(
     T const* const inputHost,
     void* const outputHost,
-    int const numBitsMax,
+    size_t const numBitsMax,
     size_t const n,
-    int* const numBitsOut,
+    size_t* const numBitsOut,
     T* const minValOut)
 {
   T* input;
@@ -186,7 +186,7 @@ void runBitPackingOnGPU(
 template<typename T>
 void typeRangeTest()
 {
-  const int numBits = 8*sizeof(T);
+  const size_t numBits = 8 * sizeof(T);
   size_t const n = 72351;
   std::vector<T> inputHost;
   inputHost.reserve(n);
@@ -202,7 +202,7 @@ void typeRangeTest()
   CUDA_RT_CALL(cudaMallocHost(&outputHost, numBytes));
 
   T minValue;
-  int numBitsAct;
+  size_t numBitsAct;
   runBitPackingOnGPU(inputHost.data(), outputHost, numBits, n, &numBitsAct, &minValue);
 
   REQUIRE(numBitsAct == numBits);
@@ -262,7 +262,7 @@ TEST_CASE("compressInt16VarBitTest", "[small]")
     }
 
     T minValueAct;
-    int numBitsAct;
+    size_t numBitsAct;
     runBitPackingOnGPU(
         inputHost, outputHost, numBits, n, &numBitsAct, &minValueAct);
 
@@ -318,7 +318,7 @@ TEST_CASE("compressUint32VarBitTest", "[small]")
     }
 
     T minValueAct;
-    int numBitsAct;
+    size_t numBitsAct;
     runBitPackingOnGPU(
         inputHost, outputHost, numBits, n, &numBitsAct, &minValueAct);
 
@@ -370,7 +370,7 @@ TEST_CASE("compressInt64VarBitTest", "[small]")
     }
 
     T minValue;
-    int numBitsAct;
+    size_t numBitsAct;
     runBitPackingOnGPU(
         inputHost, outputHost, numBits, n, &numBitsAct, &minValue);
 
@@ -397,7 +397,7 @@ TEST_CASE("compressInt64VarBitTest", "[small]")
 TEST_CASE("compressInt32VarSizeTest", "[large]")
 {
   int const offset = 87231;
-  int const numBits = 13;
+  size_t const numBits = 13;
 
   // unpack doesn't handle 0 bits
   std::vector<size_t> const sizes{2, 123, 3411, 83621, 872163, 100000001};
@@ -424,7 +424,7 @@ TEST_CASE("compressInt32VarSizeTest", "[large]")
     }
 
     T minValue;
-    int numBitsAct;
+    size_t numBitsAct;
     runBitPackingOnGPU(
         inputHost, outputHost, numBits, n, &numBitsAct, &minValue);
 
@@ -457,7 +457,7 @@ TEST_CASE("compressInt64WideTest", "[small]")
 {
   using T = int64_t;
 
-  const int numBits = 40;
+  const size_t numBits = 40;
 
   // generate a variety of random numbers
   std::vector<T> source{
@@ -479,7 +479,7 @@ TEST_CASE("compressInt64WideTest", "[small]")
   memcpy(inputHost, source.data(), sizeof(*inputHost) * source.size());
 
   T minValue;
-  int numBitsAct;
+  size_t numBitsAct;
   runBitPackingOnGPU(inputHost, outputHost, numBits, n, &numBitsAct, &minValue);
 
   REQUIRE(numBitsAct == numBits);
